Base/MailTemplateDataFixture moved to Demo/MailTemplateDataFixture
- mail templates are initialized in database migration
